requirements.txt
absl-py==1.1.0
alabaster==0.7.12
asttokens==2.0.5
attrs==21.4.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
charset-normalizer==2.0.12
cilog==1.2.3
click==8.1.3
coverage==6.4.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dgl-cu102==0.9.0
dglgo==0.0.2
docutils==0.17.1
entrypoints==0.4
et-xmlfile==1.1.0
execnet==1.9.0
executing==0.8.3
filelock==3.7.1
fonttools==4.33.3
gdown==4.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
graphviz==0.20
grpcio==1.47.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
littleutils==0.2.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
munch==2.5.0
mypy-extensions==0.4.3
nest-asyncio==1.5.5
networkx==2.8
numpy==1.23.0
numpydoc==1.4.0
oauthlib==3.2.0
ogb==1.3.3
openpyxl==3.0.10
outdated==0.2.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.9.2
PyGCL==0.1.2
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
rdkit-pypi==2022.3.3
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
torchaudio==0.11.0
torchvision==0.12.0
torchviz==0.0.2
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typed-argument-parser==1.7.2
typer==0.6.1
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
